+++ src/common/SidePanel/helpers/getAnimationDelay.ts
@@ -1,0 +1,1 @@
+export const getAnimationDelay = () => new Promise(resolve => setTimeout(resolve, 300));